# swagger.json
{
   "openapi": "3.1.0",
   "info":{
    "title": "Postaway App",
    "version": "1.0.0"
   },
   "servers":[{
    "url": "http://localhost:3200",
    "description": "Dev server"

   }],
   "components": {
    "securitySchemes": {
        "bearerAuth":{
            "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"

        }
        
    }
   },
   "paths":{
...
    "/api/posts/new": {
        "post":{
            "tags": ["Post"],
            "summary": "create new post",
            "operationId": "newPost",
            "security": [
                {
                    "bearerAuth": []

                }
                
            ],
            "parameters": [
                {
                    "name": "status",
                    "in": "query",
                    
                    "required": false,
                    "schema": {
                        "type": "string",
                        "enum" : ["draft", "published"],
                        "default": "published"
                    },
                    "description": "optional status for post"
                }
            ],
            "requestBody":{
                "required": true,
                "content":{
                    "multipart/form-data": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "caption": {
                                    "type": "string",
                                    "description": "caption of the post"
                                },
                                "image":{
                                    "type": "string",
                                    "format":"binary",
                                    "description": "image file upload"
                                }
                            },

                            "required": ["caption", "image"]

                        }
                    }
                }
            },

            "responses": {
                "201": {
                    "description": "New post successfully created",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "userId": {
                                        "type": "string"
                                    },
                                    "caption": {
                                        "type": "string"
                                    },
                                    "id": {
                                        "type":"string"
                                    },
                                    "imageUrl": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "isArchive":{
                                        "type": "boolean"
                                    }
                                }
                            }
                        }
                    }
                },

                "401": {
                    "description": "Unauthorized to create post",
                    "content": {
                        "text/plain": {
                            "schema": {
                                "type" : "string",
                                "example": "unauthorized"
                            }
                        }
                    }
                }
            }
        }
    },
...
    "/api/posts/all": {
        "get": {
            "tags": ["Post"],
            "description": "returns all posts by all users",
            "operationId": "allPosts",
            "security": [
                {
                    "bearerAuth": []
                }
            ],

            "responses":{
                "200": {
                    "description": "all posts by all users",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                "properties": {
                                    "userId": {
                                        "type": "string"
                                    },
                                    "caption": {
                                        "type": "string"
                                    },
                                    "id": {
                                        "type": "string"
                                    },
                                    "imageUrl": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "isArchive": {
                                        "type": "boolean"
                                    }
                                }
                                }
                            }
                        }
                    }
                       },

                       "401": {
                    "description": "unauthorized user",
                    "content": {
                        "text/plain": {
                            "schema": {
                                "type": "string",
                                "example": "Unauthorized"
                            }
                        }
                    }
                }
            }

        }
    },
...
    "/api/posts/post/{postId}": {
        "get": {
            "summary": "gets a single post",
            "operationId": "getSinglePost",
            "tags": ["Post"],
            "security": [

                {
                    "bearerAuth": []
                }

            ],

            "parameters": [
                {
                    "in": "path",
                    "name": "postId",
                    "required": true,
                    "schema": {
                        "type": "string"
                    }

                }
            ],

            "responses": {
                "200": {
                    "description": "returns post with given Id, if found",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    
                                    "userId": {
                                        "type": "string"
                                    },
                                    "caption": {
                                        "type": "string"
                                    },
                                    "id": {
                                        "type": "string"
                                    },
                                    "imageUrl": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "isArchive": {
                                        "type": "boolean"
                                    }
                                

                                }

                            }
                        }
                    }
                },

                "400": {
                    "description": "no post found",
                    "content": {
                        "text/plain": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },

        "post": {
            "summary": "update post",
            "operationId": "updateSinglePost",
            "tags": ["Post"],
            "security": [
                {
                    "bearerAuth": []
                }
            ],

            "parameters": [
                {
                    "in": "path",
                    "name": "postId",
                    "schema": {
                        "type": "string"
                    }
                }
            ],

            "requestBody": {
                "required": true,
                "content" :{
                    "multipart/form-data": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "caption": {
                                    "type": "string"
                                },
                                "image": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                }

            },

            "responses": {
                "201": {
                    "description": "post updated successfully",
                    "content": {
                        "text/plain": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    }
                },
                "400": {
                    "description": "no post found to update",
                    "content": {
                        "text/plain": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
...
    "/api/posts/post/delete/{postId}": {
        "delete": {
            "summary": "Delete a post",
            "operationId": "deletePost",
            "tags": ["Post"],
            "security": [
                {
                    "bearerAuth": []
                }
            ],

            "parameters": [
                {
                    "in": "path",
                    "name": "postId",
                    "schema": {
                        "type": "string"
                    }
                }
            ],

            "responses": {
                "201": {
                    "description": "post deleted successfully",
                    "content": {
                        "text/plain": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    }
                },

                "400": {
                    "description": "no post found with given id",
                    "content": {
                        "text/plain": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
...
}
}